Refactor experiment harness into shared module
